・(long)のキャスト ・UITableViewCell (Common) identifier

+++ SandBox/Controllers/ActivityViewControllerListViewController.m
@@ -99,7 +99,7 @@
     
     if (alertView.numberOfButtons){
         /* @property(nonatomic,readonly) NSInteger numberOfButtons; */
-        NSLog(@"numberOfButtons : %ld",alertView.numberOfButtons);
+        NSLog(@"numberOfButtons : %ld",(long)alertView.numberOfButtons);
     }
     
     if ([alertView respondsToSelector:@selector(buttonTitleAtIndex:)]){
@@ -111,12 +111,12 @@
     
     if (alertView.cancelButtonIndex > -1){
         /* @property(nonatomic) NSInteger cancelButtonIndex; */
-        NSLog(@"cancelButtonIndex : %ld",alertView.cancelButtonIndex);
+        NSLog(@"cancelButtonIndex : %ld",(long)alertView.cancelButtonIndex);
     }
     
     if (alertView.firstOtherButtonIndex){
         /* @property(nonatomic,readonly) NSInteger firstOtherButtonIndex; */
-        NSLog(@"firstOtherButtonIndex : %ld",alertView.firstOtherButtonIndex);
+        NSLog(@"firstOtherButtonIndex : %ld",(long)alertView.firstOtherButtonIndex);
     }
     
     if (alertView.visible){
@@ -142,23 +142,23 @@
             //            };
             
         case UIAlertViewStyleDefault : {
-            NSLog(@"alertViewStyle : %ld",UIAlertViewStyleDefault);
+            NSLog(@"alertViewStyle : %ld",(long)UIAlertViewStyleDefault);
             break;
         }
         case UIAlertViewStyleSecureTextInput : {
-            NSLog(@"alertViewStyle : %ld",UIAlertViewStyleSecureTextInput);
+            NSLog(@"alertViewStyle : %ld",(long)UIAlertViewStyleSecureTextInput);
             break;
         }
         case UIAlertViewStylePlainTextInput : {
-            NSLog(@"alertViewStyle : %ld",UIAlertViewStylePlainTextInput);
+            NSLog(@"alertViewStyle : %ld",(long)UIAlertViewStylePlainTextInput);
             break;
         }
         case UIAlertViewStyleLoginAndPasswordInput : {
-            NSLog(@"alertViewStyle : %ld",UIAlertViewStyleLoginAndPasswordInput);
+            NSLog(@"alertViewStyle : %ld",(long)UIAlertViewStyleLoginAndPasswordInput);
             break;
         }
         default:{
-            NSLog(@"alertViewStyle : %ld",alertView.alertViewStyle);
+            NSLog(@"alertViewStyle : %ld",(long)alertView.alertViewStyle);
             break;
         }
     }
